project5/requirements.txt
# SM2项目依赖包
ecdsa>=0.18.0
pycryptodome>=3.15.0
matplotlib>=3.5.0
numpy>=1.21.0
